+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+- feat: Sync annotations with line-highlight ranges when the annotation's lines are a subset of the highlight's lines.
+
 ## 1.0.0 (2026-04-17)
 
 - feat: Initial release.